restaurant-recommendations-system/ai-service/requirements.txt
fastapi
uvicorn
redis
python-dotenv
numpy
faiss-cpu
lightfm
sentence-transformers

